requirements.txt
streamlit==1.22.0
psycopg2-binary==2.9.6
pandas==1.5.3
transformers==4.29.2
torch==2.0.1
